# go-ecommerce-api/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "A RESTful E-commerce API built with Go and Gin",
        "title": "E-commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the authenticated user's shopping cart with calculated totals",
                "tags": [
                    "cart"
                ],
                "summary": "Get user's cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.CartSummaryResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove all items from the user's cart",
                "tags": [
                    "cart"
                ],
                "summary": "Clear cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cart/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a product to the user's cart",
                "tags": [
                    "cart"
                ],
                "summary": "Add item to cart",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.AddToCartRequest"
                            }
                        }
                    },
                    "description": "Add to cart request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.CartItemResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cart/items/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the quantity of an item in the cart",
                "tags": [
                    "cart"
                ],
                "summary": "Update cart item quantity",
                "parameters": [
                    {
                        "description": "Cart Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.UpdateCartItemRequest"
                            }
                        }
                    },
                    "description": "Update quantity",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove an item from the user's cart",
                "tags": [
                    "cart"
                ],
                "summary": "Remove item from cart",
                "parameters": [
                    {
                        "description": "Cart Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "models.AddToCartRequest": {
                "type": "object",
                "required": [
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 2
                    }
                }
            },
...
            "models.Cart": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.CartItem"
                        }
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "models.CartItem": {
                "type": "object",
                "properties": {
                    "cart": {
                        "$ref": "#/components/schemas/models.Cart"
                    },
                    "cart_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "models.CartItemResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/models.CartItem"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "models.CartSummary": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.CartItem"
                        }
                    },
                    "shipping": {
                        "type": "number"
                    },
                    "subtotal": {
                        "type": "number"
                    },
                    "tax": {
                        "type": "number"
                    },
                    "tax_rate": {
                        "type": "number"
                    },
                    "total": {
                        "type": "number"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "models.CartSummaryResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/models.CartSummary"
                    }
                }
            },
...
            "models.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "models.Order": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.OrderItem"
                        }
                    },
                    "payment_intent_id": {
                        "type": "string"
                    },
                    "payment_method": {
                        "description": "card, cash, etc.",
                        "type": "string"
                    },
                    "payment_status": {
                        "description": "pending, processing, succeeded, failed, cancelled",
                        "type": "string"
                    },
                    "shipping": {
                        "type": "number"
                    },
                    "status": {
                        "description": "pending, processing, shipped, delivered, cancelled",
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "transaction_id": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "models.OrderItem": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "order": {
                        "$ref": "#/components/schemas/models.Order"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "price": {
                        "description": "Price at time of order",
                        "type": "number"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "models.Product": {
                "type": "object",
                "properties": {
                    "badge": {
                        "type": "string"
                    },
                    "badge_color": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "original_price": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "models.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "message": {
                        "type": "string"
                    }
                }
            },
            "models.UpdateCartItemRequest": {
                "type": "object",
                "required": [
                    "quantity"
                ],
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 3
                    }
                }
            },
...
            "models.User": {
                "type": "object",
                "properties": {
                    "cart": {
                        "$ref": "#/components/schemas/models.Cart"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Order"
                        }
                    },
                    "role": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "wishlist": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Wishlist"
                        }
                    }
                }
            },
...
            "models.Wishlist": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "product": {
                        "$ref": "#/components/schemas/models.Product"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/models.User"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}